Search the library directory path for the standard library directory
containing lib/ and the built-in object files.  Adjust pathname->
standard-uri and standard-uri->pathname to use this directory instead
of the first directory in the path.

When plugins used --prepend-library to test a shared object, C worlds
failed.  They could not re-load registered object files, nor load
bundled options.  The guarded-system-loader only looked in the first
directory in the path, not the standard library directory.