Refactor both the stratifier and the code generator.
The stratifier now avoids the use of bit strings and just manipulates the ranges
appropriately as it groups them.  At the end it expands all the ranges so that
the nodes have minimum structure.  The code generator was modified to accept the
new input form.

The code generator has been changed to put all the terminal nodes at the
beginning of the table, and to hash-cons new non-terminal nodes.  It turns out
that there was a lot of duplication in the nodes, so this saves a bunch of
space.