Eliminate synthetic identifiers.
Now identifiers are either symbols or closures over symbols.  Any operation on a
closed identifier redirects to the appropriate environment, rather than trying
to bind and/or lookup the closure itself in the environment.

This greatly simplifies the identifier model, and makes the operation of the
syntax processor much clearer.