Extensive work to get compiler to work with R7RS libraries.
The bulk of the work is to refactor the debugging info so that it's generated
and consumed properly; this required upgrading the version and introducing a new
top-level info form.

I also eliminated the now-ancient and unused BSM file support, which was making
things more difficult than necessary.

The compiler top level required only to detect an R7RS input file and to compile
each part separately, plus collecting all of the debugging information from the
parts.  Some tweaks were used to make RTL/LAP files work right with multiple
roots.

Finally, the runtime system was modified so that compiled-code blocks and
entries now show the R7RS library name if there is one, both when printed and in
various other places.