Add a directory of pending release notes.
These are in separate files in no particular order so that we don't
have merge conflicts when people merge branches with their own
release notes.

When a release goes out we can just clear the directory and start
over.

Some of these actually went out in 10.1 but didn't make it into the
release notes (and a couple may have been slightly broken in 10.1
anyway).