Call the primitives directly.
These are no longer integrated, so we can't make a compound procedure
call if they are interpreted -- the interpreter will stomp all over
Free while we're in the middle of allocating the string, leading to
mayhem.  But it is safe to call the primitive, even though it is no
longer open-coded.